cloud_functions/handle_crawler_tasks/requirements.txt
requests
google-cloud-tasks